tools/requirements_tools.txt
pyelftools==0.32
lief==0.17.3
capstone==5.0.7
unicorn==2.1.4
